requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.1.0
matplotlib==3.7.2
seaborn==0.12.2
kafka-python==2.0.2
joblib==1.3.2
pyflink==1.17.1